Make validation more explicit
Removed ValidationConstraint with its magic accessor function
generation.

The InstanceValidation is now implemented as a method.
Also making email an explicitly required field.
